refactor: extract daily tasks rendering to DailyTasksRenderer
- Create new DailyTasksRenderer class to handle daily tasks rendering logic
- Extract renderDailyTasks method from UniversalLayoutRenderer (107 lines)
- Extract getDailyTasks method (131 lines)
- Extract getEffortArea helper (59 lines)
- Extract getChildAreas helper (28 lines)
- Extract getAssetLabel helper (51 lines)
- Reduce UniversalLayoutRenderer from 1065 to 739 lines (-327 lines, -31%)
- Maintain 100% test coverage (813 tests passing)
- Follow Clean Architecture SRP principle

Part of V-SRP-001: UniversalLayoutRenderer decomposition (Phase 2/4)
